Add files via upload
This Script is basically an detailed Text Analysis. Basically, In our dataset we had a column named 'Address'. I analyze that which pieces of string or words gives Result in positive. I create a new column 'Core_Data' in our dataframe. 
After that I develop an Pivot Table In Excel Sheet that gives 83% Positive Result.